correct dt calculation and initialized H matrix for Kalmanfilter object in FusionEKF::ProcessMeasurement; correct rho_dot calculation and also setting angle to be between -pi and pi in KalmanFilter::UpdateEKF; set all floats to double to avoid compilation warnings

+++ src/tools.cpp
@@ -20,7 +20,7 @@ VectorXd Tools::CalculateRMSE(const vector<VectorXd> &estimations,
 	}
 
 	//accumulate squared residuals
-	for(int i=0; i < estimations.size(); ++i){
+	for(double i=0; i < estimations.size(); ++i){
 		VectorXd residual = estimations[i] - ground_truth[i];
 
 		//coefficient-wise multiplication
